+++ core/src/main/java/org/apache/struts/upload/CommonsMultipartRequestHandler.java
@@ -188,10 +188,11 @@ public class CommonsMultipartRequestHandler implements MultipartRequestHandler {
             // Special handling for uploads that are too big.
             request.setAttribute(MultipartRequestHandler.ATTRIBUTE_MAX_LENGTH_EXCEEDED,
                 Boolean.TRUE);
-
+            clearInputStream(request);
             return;
         } catch (FileUploadException e) {
             log.error("Failed to parse multipart request", e);
+            clearInputStream(request);
             throw new ServletException(e);
         }
 
@@ -266,6 +267,25 @@ public class CommonsMultipartRequestHandler implements MultipartRequestHandler {
     }
 
     // -------------------------------------------------------- Support Methods
+
+    /**
+     * Finishes reading the input stream from an aborted upload. Fix for
+     * STR-2700 to prevent Window machines from hanging.
+     */
+    protected void clearInputStream(HttpServletRequest request) {
+        try {
+            if (System.getProperty("os.name").indexOf("Windows") >= -1) {
+                ServletInputStream is = request.getInputStream();
+                byte[] data = new byte[DEFAULT_SIZE_THRESHOLD];
+                int bytesRead = 0;
+                do {
+                    bytesRead = is.read(data);
+                } while (bytesRead > -1);
+            }   
+        } catch (Exception e) {
+            log.error(e.getMessage(), e);
+        }
+    }
 
     /**
      * <p> Returns the maximum allowable size, in bytes, of an uploaded file.